+++ src/helpers/dump.php
@@ -3,6 +3,7 @@
 use Codemonster\Dumper\Dumper;
 
 if (!function_exists('dump')) {
+    /** @param mixed ...$vars */
     function dump(...$vars): mixed
     {
         foreach ($vars as $var) {
@@ -14,6 +15,7 @@ if (!function_exists('dump')) {
 }
 
 if (!function_exists('dd')) {
+    /** @param mixed ...$vars */
     function dd(...$vars): never
     {
         dump(...$vars);